Refactored and extracted classes into separate py files
