+++ app.py
@@ -1,8 +1,6 @@
-import os
+from trello_request import *
 
-import requests
 from flask import Flask, render_template, request, redirect
-from requests.exceptions import HTTPError
 import dateutil.parser
 
 app = Flask(__name__, static_url_path='/static')
@@ -17,132 +15,6 @@ def format_datetime(value):
 
 
 app.jinja_env.filters['due_date_filter'] = format_datetime
-
-# Constants
-TODO_LIST_ID = os.getenv("TODO_LIST_ID")
-DONE_LIST_ID = os.getenv("DONE_LIST_ID")
-
-NOT_STARTED = 'Not Started'
-COMPLETED = 'Completed'
-
-STATUS_TO_LIST_MAP = {
-    NOT_STARTED: TODO_LIST_ID,
-    COMPLETED: DONE_LIST_ID
-}
-
-LIST_TO_STATUS_MAP = {
-    TODO_LIST_ID: NOT_STARTED,
-    DONE_LIST_ID: COMPLETED
-}
-
-
-class TrelloRequest:
-    BASE_PATH = "https://api.trello.com/1"
-    APP_API_KEY = os.getenv("APP_API_KEY")
-    APP_TOKEN = os.getenv("APP_TOKEN")
-
-    def makeRequest(self, url, method, query_params=None, **kwargs):
-        requestUrl = self.BASE_PATH + url
-        app.logger.info(f"Trello {method} request {requestUrl} with queryParams: {query_params} and additional request: {kwargs}")
-
-        try:
-            query = {
-                'key': self.APP_API_KEY,
-                'token': self.APP_TOKEN
-            }
-
-            if query_params:
-                query.update(query_params)
-
-            response = requests.request(
-                method,
-                requestUrl,
-                params=query,
-                **kwargs
-            )
-
-            response.raise_for_status()
-
-            jsonResponse = response.json()
-
-            app.logger.info(jsonResponse)
-
-            return jsonResponse
-
-        except HTTPError as http_err:
-            app.logger.error(f'HTTP error occurred: {http_err}')
-        except Exception as err:
-            app.logger.error(f'Other error occurred: {err}')
-        raise Exception("Request failed. See logs.")
-
-
-class TrelloGetCards(TrelloRequest):
-    URL_PATH = "/lists/{}/cards"
-
-    def fetchForList(self, idList):
-        url = self.URL_PATH.format(idList)
-        return super().makeRequest(url, "GET")
-
-    def fetchCard(self, id):
-        url = "/card/{}".format(id)
-
-        jsonData = super().makeRequest(url, "GET")
-
-        return TodoItem(jsonData['name'], LIST_TO_STATUS_MAP[jsonData["idList"]], id)
-
-
-class TrelloAddCard(TrelloRequest):
-    URL_PATH = "/cards"
-
-    def __init__(self, listId):
-        self.listId = listId
-
-    def add(self, item):
-        url = self.URL_PATH
-
-        query = {
-            'idList': self.listId,
-            'name': item.title
-        }
-
-        if item.duedate:
-            query.update({"due": item.duedate})
-
-        return super().makeRequest(url, "POST", query)
-
-
-class TrelloUpdateCard(TrelloRequest):
-    URL_PATH = "/cards/{}"
-
-    def update(self, item, listId):
-        url = self.URL_PATH.format(item.id)
-
-        headers = {
-            "Accept": "application/json"
-        }
-
-        query = {
-            'name': item.title,
-            'idList': listId
-        }
-
-        return super().makeRequest(url, "PUT", query, headers=headers)
-
-
-class TrelloDeleteCard(TrelloRequest):
-    URL_PATH = "/cards/{}"
-
-    def delete(self, cardId):
-        url = self.URL_PATH.format(cardId)
-        return super().makeRequest(url, "DELETE")
-
-
-class TodoItem:
-    def __init__(self, title, status, id=None, duedate=None):
-        self.id = id
-        self.title = title
-        self.status = status
-        self.duedate = duedate
 
 
 @app.route('/')
